+++ gluoncv/data/transforms/presets/center_net.py
@@ -69,40 +69,6 @@ def _affine_transform(pt, t):
     new_pt = np.dot(t, new_pt)
     return new_pt[:2]
 
-def grayscale(image):
-    cv2 = try_import_cv2()
-    return cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-
-def lighting_(data_rng, image, alphastd, eigval, eigvec):
-    alpha = data_rng.normal(scale=alphastd, size=(3, ))
-    image += np.dot(eigvec, eigval * alpha)
-
-def blend_(alpha, image1, image2):
-    image1 *= alpha
-    image2 *= (1 - alpha)
-    image1 += image2
-
-def saturation_(data_rng, image, gs, gs_mean, var):
-    alpha = 1. + data_rng.uniform(low=-var, high=var)
-    blend_(alpha, image, gs[:, :, None])
-
-def brightness_(data_rng, image, gs, gs_mean, var):
-    alpha = 1. + data_rng.uniform(low=-var, high=var)
-    image *= alpha
-
-def contrast_(data_rng, image, gs, gs_mean, var):
-    alpha = 1. + data_rng.uniform(low=-var, high=var)
-    blend_(alpha, image, gs_mean)
-
-def color_aug(data_rng, image, eig_val, eig_vec):
-    functions = [brightness_, contrast_, saturation_]
-    random.shuffle(functions)
-
-    gs = grayscale(image)
-    gs_mean = gs.mean()
-    for f in functions:
-        f(data_rng, image, gs, gs_mean, 0.4)
-    lighting_(data_rng, image, 0.1, eig_val, eig_vec)
 
 class CenterNetDefaultTrainTransform(object):
     """Default SSD training transform which includes tons of image augmentations.
@@ -179,7 +145,7 @@ class CenterNetDefaultTrainTransform(object):
 
         # to tensor
         img = img.astype(np.float32) / 255.
-        color_aug(self._data_rng, img, self._eig_val, self._eig_vec)
+        experimental.np_random_color_distort(img, data_rng=self._data_rng)
         img = (img - self._mean) / self._std
         img = img.transpose(2, 0, 1).astype(np.float32)
         img = mx.nd.array(img)
@@ -285,7 +251,7 @@ class CenterNetDefaultTrainTransformDebug(object):
 
         # to tensor
         img = img.astype(np.float32) / 255.
-        color_aug(self._data_rng, img, self._eig_val, self._eig_vec)
+        experimental.np_random_color_distort(img, data_rng=self._data_rng)
         img = (img - self._mean) / self._std
         img = img.transpose(2, 0, 1).astype(np.float32)
         # img = mx.nd.image.to_tensor(img)
